// frontend/src/openapi/generated/model/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zete
 * 메모 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Category } from './category';
// May contain unused imports in some cases
// @ts-ignore
import { Memo } from './memo';
// May contain unused imports in some cases
// @ts-ignore
import { Tag } from './tag';

/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * 
     * @type {number}
     * @memberof User
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'createAt': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'updateAt': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'email': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'password': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'mobile': string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'refreshToken'?: string;
    /**
     * 
     * @type {number}
     * @memberof User
     */
    'gptUsableCount'?: number;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'gptUsableCountResetAt'?: string;
    /**
     * 
     * @type {Array<Category>}
     * @memberof User
     */
    'cates'?: Array<Category>;
    /**
     * 
     * @type {Array<Memo>}
     * @memberof User
     */
    'memos'?: Array<Memo>;
    /**
     * 
     * @type {Array<Tag>}
     * @memberof User
     */
    'tags'?: Array<Tag>;
}